Move fix_doc.py under package
